requirements.txt
pyusb==1.2.1
colorlog==6.8.2